# Pertemuan 4/project_mysql/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TokoBersama API",
    "version": "1.0.0",
    "description": "REST API untuk platform marketplace TokoBersama. Mengelola kategori produk, seller, produk, dan order."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "components": {
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string" },
          "data": {}
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" },
          "data": { "nullable": true, "example": null }
        }
      },
...
      "Seller": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "nama_toko": { "type": "string", "example": "TechZone Official" },
          "email": { "type": "string", "example": "techzone@mail.com" },
          "no_hp": { "type": "string", "example": "081234567890" },
          "alamat": {
            "type": "string",
            "example": "Jl. Sudirman No. 10, Jakarta"
          },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "SellerInput": {
        "type": "object",
        "required": ["nama_toko", "email"],
        "properties": {
          "nama_toko": { "type": "string", "example": "TechZone Official" },
          "email": { "type": "string", "example": "techzone@mail.com" },
          "no_hp": { "type": "string", "example": "081234567890" },
          "alamat": {
            "type": "string",
            "example": "Jl. Sudirman No. 10, Jakarta"
          }
        }
      },
      "Produk": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "nama": { "type": "string", "example": "Samsung Galaxy A55" },
          "harga": { "type": "number", "example": 4999000 },
          "stok": { "type": "integer", "example": 50 },
          "id_kategori": { "type": "integer", "example": 1 },
          "id_seller": { "type": "integer", "example": 1 },
          "nama_kategori": { "type": "string", "example": "Elektronik" },
          "nama_toko": { "type": "string", "example": "TechZone Official" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "ProdukInput": {
        "type": "object",
        "required": ["nama", "harga", "id_kategori", "id_seller"],
        "properties": {
          "nama": { "type": "string", "example": "Samsung Galaxy A55" },
          "harga": { "type": "number", "example": 4999000 },
          "stok": { "type": "integer", "example": 50 },
          "id_kategori": { "type": "integer", "example": 1 },
          "id_seller": { "type": "integer", "example": 1 }
        }
      },
...
    }
  },
  "paths": {
...
    "/seller": {
      "get": {
        "tags": ["Seller"],
        "summary": "Ambil semua seller",
        "responses": {
          "200": {
            "description": "Daftar semua seller",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Seller" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Seller"],
        "summary": "Daftarkan seller baru",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SellerInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Seller berhasil didaftarkan",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Seller" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validasi gagal / email duplikat",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/seller/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer" },
          "example": 1
        }
      ],
      "get": {
        "tags": ["Seller"],
        "summary": "Ambil seller berdasarkan ID",
        "responses": {
          "200": {
            "description": "Data seller",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Seller" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Seller tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Seller"],
        "summary": "Update data seller",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SellerInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Seller berhasil diupdate",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Seller" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validasi gagal / email duplikat",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Seller tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Seller"],
        "summary": "Hapus seller",
        "responses": {
          "200": {
            "description": "Seller berhasil dihapus",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "404": {
            "description": "Seller tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/seller/{id}/produk": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer" },
          "example": 1
        }
      ],
      "get": {
        "tags": ["Seller"],
        "summary": "Ambil semua produk milik seller tertentu",
        "responses": {
          "200": {
            "description": "Daftar produk milik seller",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Produk" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Seller tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/produk": {
      "get": {
        "tags": ["Produk"],
        "summary": "Ambil semua produk",
        "description": "Mendukung filter `?kategori=:id` dan pencarian `?search=:keyword`. Hasil sudah termasuk nama_kategori dan nama_toko (JOIN).",
        "parameters": [
          {
            "name": "kategori",
            "in": "query",
            "required": false,
            "schema": { "type": "integer" },
            "description": "Filter produk berdasarkan ID kategori",
            "example": 1
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Cari produk berdasarkan nama (LIKE)",
            "example": "samsung"
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar produk",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Produk" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Produk"],
        "summary": "Tambah produk baru",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProdukInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produk berhasil ditambahkan",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Produk" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validasi gagal / FK tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/produk/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer" },
          "example": 1
        }
      ],
      "get": {
        "tags": ["Produk"],
        "summary": "Ambil detail produk berdasarkan ID",
        "responses": {
          "200": {
            "description": "Detail produk",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Produk" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Produk tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Produk"],
        "summary": "Update produk (termasuk stok)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProdukInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Produk berhasil diupdate",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Produk" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validasi gagal",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Produk tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Produk"],
        "summary": "Hapus produk",
        "responses": {
          "200": {
            "description": "Produk berhasil dihapus",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "404": {
            "description": "Produk tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}